add 3 turret
i forgot to  pull up

+++ Assets/Content/Features/Turret/TuretModule/TurretAIRottate.cs
@@ -7,8 +7,8 @@ public class TurretAIRottate : MonoBehaviour
     
     Transform[] target;
     
-    [Range(0, 100f)] public float MaxDistance;
-    [Range(0, 100f)] public float MinDistance;
+    [Range(0, 100)] public int MaxDistance;
+    [Range(0, 100)] public int MinDistance;
 
     private void Awake()
     {
